requirements.txt
amqtt
asyncio
mysql-connector-python
cryptography
elasticsearch
pandas